openai-mcp-client/requirements.txt
openai>=1.68.0
python-dotenv>=1.0.0
mcp[cli]
asyncio>=3.4.3

